requirements.txt
Flask==3.0.0
numpy==1.24.3
Pillow==10.1.0
joblib==1.3.2
scikit-image==0.21.0
scikit-learn==1.3.2

